requirements.txt
shapely>=2.0.1
python-geohash>=0.8.5
setuptools>=67.6.0
